Investigate panic density and introduce fuzz harnesses for untrusted parsers
Extract tensor protobuf/JSON decoding into sn2-types::tensor_codec and the
length-prefixed frame codec into a new sn2-frame-codec crate so the parsers
sit on a minimal dependency footprint. Replace remaining production-code
panic and unwrap sites in validator_loop::dslice and miner::dsperse with
graceful error returns. Add a libfuzzer harness covering gzipped protobuf
tensor decode, JSON tensor flattening, single-shot frame reads, and
chunked-stream frame reads, plus a CI workflow that smoke-runs each target
under address sanitizer.
