+++ crates/sn2-types/Cargo.toml
@@ -9,3 +9,5 @@ anyhow = { workspace = true }
 serde = { workspace = true }
 serde_json = { workspace = true }
 tracing-subscriber = { workspace = true }
+flate2 = { workspace = true }
+ndarray = { workspace = true }
